pyproject.toml
[tool.poetry]
name = "cmsc125-mp1"
version = "0.1.0"
description = ""
authors = ["melzarjanchico <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
py-getch = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
